Create get-a-diff-of-the-dependencies-between-commits.sh
Add get-a-diff-of-the-dependencies-between-commits.sh which is used to get the diff of the dependency changes between two commits of a repository, based on the changes to the dependency manifests made in those commits. If no variables are provided, query commits API to compare the first and last commits in the response.